remove boost::geometry from the interface completely
