+++ src/mapping/RadialBasisFctMapping.hpp
@@ -5,18 +5,10 @@
 #include <Eigen/Core>
 #include <Eigen/QR>
 
-#include <boost/version.hpp>
-#if BOOST_VERSION < 106600
-#include <boost/function_output_iterator.hpp>
-#else
-#include <boost/iterator/function_output_iterator.hpp>
-#endif
-
 #include "com/CommunicateMesh.hpp"
 #include "com/Communication.hpp"
 #include "impl/BasisFunctions.hpp"
 #include "mesh/Filter.hpp"
-#include "mesh/impl/BBUtils.hpp"
 #include "query/RTree.hpp"
 #include "utils/EigenHelperFunctions.hpp"
 #include "utils/Event.hpp"
