+++ src/query/RTree.cpp
@@ -1,12 +1,23 @@
 #include "query/RTree.hpp"
+#include <Eigen/Core>
 #include <algorithm>
+#include <boost/geometry.hpp>
 #include <boost/range/irange.hpp>
 #include <cstddef>
 #include <utility>
 #include <vector>
 #include "math/barycenter.hpp"
 #include "mesh/Vertex.hpp"
+#include "query/RTreeTools.hpp"
+#include "query/impl/RTreeAdapter.hpp"
 #include "utils/assertion.hpp"
+
+#include <boost/version.hpp>
+#if BOOST_VERSION < 106600
+#include <boost/function_output_iterator.hpp>
+#else
+#include <boost/iterator/function_output_iterator.hpp>
+#endif
 
 namespace precice {
 namespace query {
@@ -15,84 +26,11 @@ namespace bg  = boost::geometry;
 namespace bgi = boost::geometry::index;
 
 // Initialize static member
-std::map<int, rtree::MeshIndices> precice::query::rtree::_cachedTrees;
-
-rtree::MeshIndices &rtree::cacheEntry(int meshID)
-{
-  auto result = _cachedTrees.emplace(std::make_pair(meshID, rtree::MeshIndices{}));
-  return result.first->second;
-}
-
-VertexTraits::Ptr rtree::getVertexRTree(const mesh::PtrMesh &mesh)
-{
-  PRECICE_ASSERT(mesh);
-  auto &cache = cacheEntry(mesh->getID());
-  if (cache.vertices) {
-    return cache.vertices;
-  }
-
-  // Generating the rtree is expensive, so passing everything in the ctor is
-  // the best we can do. Even passing an index range instead of calling
-  // tree->insert repeatedly is about 10x faster.
-  RTreeParameters           params;
-  VertexTraits::IndexGetter ind(mesh->vertices());
-  auto                      tree = std::make_shared<VertexTraits::RTree>(
-      boost::irange<std::size_t>(0lu, mesh->vertices().size()), params, ind);
-
-  cache.vertices = tree;
-  return tree;
-}
-
-EdgeTraits::Ptr rtree::getEdgeRTree(const mesh::PtrMesh &mesh)
-{
-  PRECICE_ASSERT(mesh);
-  auto &cache = cacheEntry(mesh->getID());
-  if (cache.edges) {
-    return cache.edges;
-  }
-
-  // Generating the rtree is expensive, so passing everything in the ctor is
-  // the best we can do. Even passing an index range instead of calling
-  // tree->insert repeatedly is about 10x faster.
-  RTreeParameters         params;
-  EdgeTraits::IndexGetter ind(mesh->edges());
-  auto                    tree = std::make_shared<EdgeTraits::RTree>(
-      boost::irange<std::size_t>(0lu, mesh->edges().size()), params, ind);
-
-  cache.edges = tree;
-  return tree;
-}
-
-TriangleTraits::Ptr rtree::getTriangleRTree(const mesh::PtrMesh &mesh)
-{
-  PRECICE_ASSERT(mesh);
-  auto &cache = cacheEntry(mesh->getID());
-  if (cache.triangles) {
-    return cache.triangles;
-  }
-
-  // We first generate the values for the triangle rtree.
-  // The resulting vector is a random access range, which can be passed to the
-  // constructor of the rtree for more efficient indexing.
-  std::vector<TriangleTraits::IndexType> elements;
-  elements.reserve(mesh->triangles().size());
-  for (size_t i = 0; i < mesh->triangles().size(); ++i) {
-    auto box = bg::return_envelope<RTreeBox>(mesh->triangles()[i]);
-    elements.emplace_back(std::move(box), i);
-  }
-
-  // Generating the rtree is expensive, so passing everything in the ctor is
-  // the best we can do.
-  RTreeParameters             params;
-  TriangleTraits::IndexGetter ind;
-  auto                        tree = std::make_shared<TriangleTraits::RTree>(elements, params, ind);
-  cache.triangles                  = tree;
-  return tree;
-}
+std::map<int, MeshIndices> precice::query::RTreeTools::_cachedTrees;
 
 std::vector<MatchType> rtree::getClosestVertex(const mesh::Vertex &source, const mesh::PtrMesh &targetMesh, int n)
 {
-  auto                   tree = getVertexRTree(targetMesh);
+  auto                   tree = RTreeTools::getVertexRTree(targetMesh);
   std::vector<MatchType> matches;
   tree->query(bgi::nearest(source, n), boost::make_function_output_iterator([&](size_t matchID) {
                 matches.emplace_back(boost::geometry::distance(source, targetMesh->vertices()[matchID]), matchID);
@@ -103,7 +41,7 @@ std::vector<MatchType> rtree::getClosestVertex(const mesh::Vertex &source, const
 
 std::vector<MatchType> rtree::getClosestEdge(const mesh::Vertex &source, const mesh::PtrMesh &targetMesh, int n)
 {
-  auto                   tree = getEdgeRTree(targetMesh);
+  auto                   tree = RTreeTools::getEdgeRTree(targetMesh);
   std::vector<MatchType> matches;
   tree->query(bgi::nearest(source, n), boost::make_function_output_iterator([&](size_t matchID) {
                 matches.emplace_back(boost::geometry::distance(source, targetMesh->edges()[matchID]), matchID);
@@ -114,7 +52,7 @@ std::vector<MatchType> rtree::getClosestEdge(const mesh::Vertex &source, const m
 
 std::vector<MatchType> rtree::getClosestTriangle(const mesh::Vertex &source, const mesh::PtrMesh &targetMesh, int n)
 {
-  auto                   tree = getTriangleRTree(targetMesh);
+  auto                   tree = RTreeTools::getTriangleRTree(targetMesh);
   std::vector<MatchType> matches;
   tree->query(bgi::nearest(source, n),
               boost::make_function_output_iterator([&](TriangleTraits::IndexType const &match) {
@@ -124,9 +62,10 @@ std::vector<MatchType> rtree::getClosestTriangle(const mesh::Vertex &source, con
   return matches;
 }
 
-std::vector<size_t> rtree::getVerticesInsideBox(const Box3d &searchBox, const mesh::PtrMesh &targetMesh, const mesh::Vertex &centerVertex, double supportRadius)
+std::vector<size_t> rtree::getVerticesInsideBox(const mesh::PtrMesh &targetMesh, const mesh::Vertex &centerVertex, double supportRadius)
 {
-  auto                tree = getVertexRTree(targetMesh);
+  auto                searchBox = RTreeTools::getEnclosingBox(centerVertex, supportRadius);
+  auto                tree      = RTreeTools::getVertexRTree(targetMesh);
   std::vector<size_t> matches;
   tree->query(bgi::intersects(searchBox) and bg::index::satisfies([&](size_t const i) { return bg::distance(centerVertex, targetMesh->vertices()[i]) <= supportRadius; }),
               std::back_inserter(matches));
@@ -135,7 +74,7 @@ std::vector<size_t> rtree::getVerticesInsideBox(const Box3d &searchBox, const me
 
 void rtree::tagAllInsideBox(const mesh::BoundingBox &boundingBox, const mesh::PtrMesh &targetMesh)
 {
-  auto tree = getVertexRTree(targetMesh);
+  auto tree = RTreeTools::getVertexRTree(targetMesh);
   tree->query(bgi::intersects(RTreeBox(boundingBox.minCorner(), boundingBox.maxCorner())),
               boost::make_function_output_iterator([&targetMesh](size_t idx) {
                 targetMesh->vertices()[idx].tag();
@@ -144,29 +83,12 @@ void rtree::tagAllInsideBox(const mesh::BoundingBox &boundingBox, const mesh::Pt
 
 void rtree::clear(mesh::Mesh &mesh)
 {
-  _cachedTrees.erase(mesh.getID());
+  RTreeTools::clear(mesh);
 }
 
 void rtree::clear()
 {
-  _cachedTrees.clear();
-}
-
-Box3d getEnclosingBox(mesh::Vertex const &middlePoint, double sphereRadius)
-{
-  namespace bg = boost::geometry;
-  auto &coords = middlePoint.getCoords();
-
-  Box3d box;
-  bg::set<bg::min_corner, 0>(box, bg::get<0>(coords) - sphereRadius);
-  bg::set<bg::min_corner, 1>(box, bg::get<1>(coords) - sphereRadius);
-  bg::set<bg::min_corner, 2>(box, bg::get<2>(coords) - sphereRadius);
-
-  bg::set<bg::max_corner, 0>(box, bg::get<0>(coords) + sphereRadius);
-  bg::set<bg::max_corner, 1>(box, bg::get<1>(coords) + sphereRadius);
-  bg::set<bg::max_corner, 2>(box, bg::get<2>(coords) + sphereRadius);
-
-  return box;
+  RTreeTools::clear();
 }
 
 } // namespace query
